# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Apoia API",
    "version": "1.0"
  },
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWE"
      }
    }
  },
  "paths": {
    "/api/v1/ai": {
      "post": {
        "summary": "Executa um prompt de IA e retorna a resposta (stream ou texto)",
        "description": "Gera uma resposta a partir de diversos parâmetros de configuração de prompt e dados obtidos de um processo ou texto arbitrário.",
        "tags": [
          "ai"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "kind",
                  "data"
                ],
                "properties": {
                  "kind": {
                    "type": "string",
                    "description": "Identificador do prompt interno ou personalizado (ex: chat, resumo, prompt-123)."
                  },
                  "promptSlug": {
                    "type": "string",
                    "description": "Slug de um prompt oficial para o tipo indicado em kind."
                  },
                  "promptId": {
                    "type": "integer",
                    "description": "ID numérico de um prompt salvo."
                  },
                  "data": {
                    "type": "object",
                    "description": "Dados de entrada (ex: numeroDoProcesso, textos, etc) dependentes do prompt."
                  },
                  "overrideSystemPrompt": {
                    "type": "string",
                    "description": "Sobrescreve o system prompt."
                  },
                  "overridePrompt": {
                    "type": "string",
                    "description": "Sobrescreve o prompt principal."
                  },
                  "overrideJsonSchema": {
                    "type": "string",
                    "description": "Força um JSON Schema para validação da saída."
                  },
                  "overrideFormat": {
                    "type": "string",
                    "description": "Formato alternativo (ex: MARKDOWN, HTML, JSON_LINES)."
                  },
                  "overrideTemplate": {
                    "type": "string",
                    "description": "Template de composição (caso use prompts internos com placeholders)."
                  },
                  "cacheControl": {
                    "type": "boolean",
                    "description": "Ativa/desativa mecanismo de cache da geração."
                  },
                  "modelSlug": {
                    "type": "string",
                    "description": "Modelo de IA a ser usado (sobrescreve configuração padrão)."
                  },
                  "extra": {
                    "type": "string",
                    "description": "Texto adicional concatenado ao prompt final."
                  },
                  "dossierCode": {
                    "type": "string",
                    "description": "Código do dossiê / processo usado para auditoria."
                  },
                  "documentId": {
                    "type": "string",
                    "description": "Identificador de documento associado (quando aplicável)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resposta do assistente (texto plano ou JSON se jsonSchema for utilizado).",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "Conteúdo textual serializado ou objeto conforme schema solicitado."
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida (parâmetros incorretos / prompt inexistente)."
          },
          "401": {
            "description": "Não autorizado."
          },
          "405": {
            "description": "Erro durante a execução do prompt ou comunicação com o provedor."
          }
        }
      }
    },
...
  }
}